Przychód liczony z istnienia dokumentu, koszty w groszach, koniec z backfillem
Trzy zmiany, które mają jedną wspólną przyczynę: sumy pieniężne wyliczały
stany zamiast wyrażać regułę, a kwoty kosztowe żyły na liczbach
zmiennoprzecinkowych.

1. Faktura niewysłana do KSeF to pełnoprawny przychód.
   findTotalStatistics, findMonthlyStatistics i sumGrossByPaymentStatus
   filtrowały po białej liście statusów wysyłki. NOT_SENT, dodany razem
   z przełącznikiem wysyłki, nie trafił na tę listę i sprzedaż znikała
   z obu widoków — a powiązany dokument finansowy i tak jest pomijany
   przez ksef_revenue_invoice_id, więc nie liczyła się nigdzie. Biała
   lista ustępuje czarnej: odpada tylko REJECTED, czyli dokument, którego
   KSeF nie przyjął. Kolejny status cyklu wysyłki będzie się domyślnie
   liczył, zamiast po cichu wypadać.

   Przy okazji ta sama klasa błędu po stronie dokumentów finansowych:
   raport sumował wyłącznie PAID i PENDING, więc dokument OVERDUE był
   wart zero złotych zarówno w przychodzie, jak i w należnościach.
   sumGross przyjmuje teraz zbiór statusów, a należność przeterminowana
   sumuje się razem z oczekującą.

2. Kwoty faktur kosztowych w groszach (BIGINT), jak po stronie
   przychodowej. Dotąd były to złote na DOUBLE PRECISION: raport mnożył
   sumę przez 100, a porównanie kwoty „co do grosza" — potrzebne do
   wykrywania zdublowanych faktur — było na floatach niewykonalne.
   Migracja V80 konwertuje przez NUMERIC, więc 45,45 zł daje dokładnie
   4545 groszy. Kontrakt HTTP pozostaje w złotych; konwersja siedzi
   wyłącznie na granicy kontrolerów.

3. Nie ma już pobierania kosztów z dowolnego zakresu dat. Endpoint
   POST /v1/ksef/expenses/sync przyjmował dateFrom i dateTo prosto z ciała
   żądania i przekazywał je do handlera bez żadnego ograniczenia, więc
   każdy manager mógł zaciągnąć historię sprzed integracji i zdublować
   koszty wprowadzone wcześniej ręcznie. Zakres ustala teraz wyłącznie
   KsefSyncService, który nigdy nie schodzi poniżej momentu podania
   tokenu — tak samo jak po stronie przychodów.

+++ src/main/kotlin/pl/detailing/crm/ksef/infrastructure/KsefInvoiceEntity.kt
@@ -95,14 +95,15 @@ class KsefInvoiceEntity(
     @Column(name = "buyer_country_code", length = 2)
     val buyerCountryCode: String? = null,
 
+    /** Kwoty w GROSZACH, jak po stronie przychodowej. Patrz V80. */
     @Column(name = "net_amount")
-    val netAmount: Double?,
+    val netAmount: Long?,
 
     @Column(name = "gross_amount")
-    val grossAmount: Double?,
+    val grossAmount: Long?,
 
     @Column(name = "vat_amount")
-    val vatAmount: Double?,
+    val vatAmount: Long?,
 
     @Column(name = "currency", length = 3)
     val currency: String?,
@@ -224,9 +225,9 @@ class KsefInvoiceEntity(
         sellerName: String? = this.sellerName,
         buyerNip: String? = this.buyerNip,
         buyerName: String? = this.buyerName,
-        netAmount: Double? = this.netAmount,
-        grossAmount: Double? = this.grossAmount,
-        vatAmount: Double? = this.vatAmount,
+        netAmount: Long? = this.netAmount,
+        grossAmount: Long? = this.grossAmount,
+        vatAmount: Long? = this.vatAmount,
         currency: String? = this.currency,
         invoiceType: String? = this.invoiceType,
         fetchedAt: Instant = this.fetchedAt,
